refactor(pipelines): trim the public surface and drop dead result fields
Stop exporting the engine's internals (finishResult, toRunInputs, the
standalone output helpers) from the package index; callers reach the
helpers through pipelineOutputs / context.outputs. Remove the fields and
event variants nothing ever set or emitted: viewUrl and summary on step
results and graph nodes, the block_output and agent_event events, the
'branch' control kind and the node-definition version. registerNode
already validates ids, so the second validateNodeId call goes too.

RunBlockOutput moves into extract-outputs.ts; block-output.ts only
existed to mirror runtime-core's AgentStreamEvent and ExecutionSummary
for the removed fields, so it and the local-runner test asserting that
mirror are deleted.
