+++ packages/pipelines/src/index.ts
@@ -1,6 +1,6 @@
-export type { AgentStreamEvent, ExecutionSummary, RunBlockOutput } from './block-output'
 export type { CloudExecutorOptions } from './cloud-executor'
-export { createCloudStepExecutor, DEFAULT_CLOUD_API_URL, toRunInputs } from './cloud-executor'
+export { createCloudStepExecutor, DEFAULT_CLOUD_API_URL } from './cloud-executor'
+export type { RunBlockOutput } from './extract-outputs'
 export { extractOutputs } from './extract-outputs'
 export type { InputBlockInfo } from './input-info'
 export { inputInfoFor } from './input-info'
@@ -24,17 +24,6 @@ export type {
   PipelineStepExecutor,
   PipelineStepResult,
 } from './pipeline'
-export {
-  allOutputText,
-  finishResult,
-  lastAgentText,
-  lastOutputJson,
-  outputJson,
-  outputText,
-  PipelineStepError,
-  pipelineOutputs,
-  runPipeline,
-  runPipelineWithExecutor,
-} from './pipeline'
+export { PipelineStepError, pipelineOutputs, runPipeline, runPipelineWithExecutor } from './pipeline'
 export type { SnapshotBlock, SnapshotInput, SnapshotNotebook, SnapshotView } from './snapshot-view'
 export { parseSnapshot, toSnapshotView } from './snapshot-view'
